+++ packages/desktop/components/drawers/network-config/drawers/components/index.ts
@@ -1,0 +1,1 @@
+export { default as AddIscpChainForm } from './AddIscpChainForm.svelte'
