+++ packages/shared/lib/core/utils/crypto/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './isValidHexAddress'
 export * from './validateAssetId'
 export * from './validateBech32Address'
 export * from './validateEthereumAddress'
