+++ packages/shared/locales/en.json
@@ -495,7 +495,14 @@
                         "chainId": "Chain ID",
                         "aliasAddress": "Alias Address",
                         "iscpEndpoint": "ISCP Endpoint",
-                        "explorerEndpoint": "Explorer Endpoint (optional)"
+                        "explorerEndpoint": "Explorer Endpoint (optional)",
+                        "errors": {
+                            "cannotBeEmpty": "Name cannot be empty",
+                            "nameTooLong": "Name is too long",
+                            "aliasAddressMustBeHex": "Alias address must be in hex format",
+                            "invalidUrl": "Invalid URL"
+                        }
+                        
                     },
                     "chainDepositAddress": {
                         "title": "Chain deposit address"
